+++ nglview/theme/jsutils.py
@@ -1,0 +1,9 @@
+js_clean_empty_output_area = """
+var output_area = $(".output_area");
+
+for (var i=0; i < output_area.length; i++){
+    if (!output_area[i].innerText){
+        output_area[i].remove();
+    }
+}
+"""